rag-product/requirements.txt
google-generativeai>=0.3.0
streamlit>=1.32.0
python-dotenv>=1.0.0
